+++ backend/backend.go
@@ -77,7 +77,6 @@ var fixedURLWhitelist = []string{
 	"https://sochain.com/tx/LTCTEST/",
 	"https://blockchair.com/litecoin/transaction/",
 	"https://etherscan.io/tx/",
-	"https://goerli.etherscan.io/tx/",
 	"https://sepolia.etherscan.io/tx/",
 	// Moonpay onramp
 	"https://www.moonpay.com/",
@@ -509,12 +508,6 @@ func (backend *Backend) Coin(code coinpkg.Code) (coinpkg.Coin, error) {
 		etherScan := etherscan.NewEtherScan("https://api.etherscan.io/api", backend.etherScanHTTPClient)
 		coin = eth.NewCoin(etherScan, code, "Ethereum", "ETH", "ETH", params.MainnetChainConfig,
 			"https://etherscan.io/tx/",
-			etherScan,
-			nil)
-	case code == coinpkg.CodeGOETH:
-		etherScan := etherscan.NewEtherScan("https://api-goerli.etherscan.io/api", backend.etherScanHTTPClient)
-		coin = eth.NewCoin(etherScan, code, "Ethereum Goerli", "GOETH", "GOETH", params.GoerliChainConfig,
-			"https://goerli.etherscan.io/tx/",
 			etherScan,
 			nil)
 	case code == coinpkg.CodeSEPETH: